metrics: enable Compose OTEL experiment
Docker Desktop has a feature flag that will slowly roll out the
OpenTelemetry collector that can export events from the CLI to
Segment.

When that feature flag is set, it adds an `otel` section to the
Docker context metadata. Within `compose-cli`, if we see that
after loading the context, set `COMPOSE_EXPERIMENTAL_OTEL` on
the process so that it gets passed to the subprocess, aka the
actual `docker` CLI, which will then call the `docker-compose`
plugin binary.

This prevents needing extra logic here to read from either the
Docker Desktop configuration (which is a no go) or the kind-of
related `features.json`, which is used by `compose-switch` for
v1/v2 handling. Most importantly, it prevents more file I/O,
because this binary already reads too many files from disk,
which then get immediately re-read by the other CLIs that it
invokes.

Long-term, we should figure out a way to set experimental env
vars for our CLIs that we want to enable in Docker Desktop in
a consistent and low-impact way. In particular, `compose-cli`
has been deprecated, and we're hoping to eliminate the shim
process approach entirely in the not-too-distant future.

Signed-off-by: Milas Bowman <milas.bowman@docker.com>